backend/users/serializers.py
from rest_framework import serializers

from . import models


class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.CustomUser
        fields = ("id", "username")
